add and consistently use parseDate and renderDateBadge helpers
also move

- age
- formatDate
- formatRelativeDate

to date.js
