+++ services/chrome-web-store/chrome-web-store-last-updated.service.js
@@ -1,5 +1,4 @@
-import { age } from '../color-formatters.js'
-import { formatDate } from '../text-formatters.js'
+import { renderDateBadge } from '../date.js'
 import { NotFound, pathParams } from '../index.js'
 import BaseChromeWebStoreService from './chrome-web-store-base.js'
 
@@ -31,11 +30,6 @@ export default class ChromeWebStoreLastUpdated extends BaseChromeWebStoreService
       throw new NotFound({ prettyMessage: 'not found' })
     }
 
-    const lastUpdatedDate = Date.parse(lastUpdated)
-
-    return {
-      message: formatDate(lastUpdatedDate),
-      color: age(lastUpdatedDate),
-    }
+    return renderDateBadge(lastUpdated)
   }
 }
