+++ services/eclipse-marketplace/eclipse-marketplace-update.service.js
@@ -1,7 +1,6 @@
 import Joi from 'joi'
 import { pathParams } from '../index.js'
-import { formatDate } from '../text-formatters.js'
-import { age as ageColor } from '../color-formatters.js'
+import { renderDateBadge } from '../date.js'
 import { nonNegativeInteger } from '../validators.js'
 import EclipseMarketplaceBase from './eclipse-marketplace-base.js'
 
@@ -34,19 +33,12 @@ export default class EclipseMarketplaceUpdate extends EclipseMarketplaceBase {
 
   static defaultBadgeData = { label: 'updated' }
 
-  static render({ date }) {
-    return {
-      message: formatDate(date),
-      color: ageColor(date),
-    }
-  }
-
   async handle({ name }) {
     const { marketplace } = await this.fetch({
       name,
       schema: updateResponseSchema,
     })
     const date = 1000 * parseInt(marketplace.node.changed)
-    return this.constructor.render({ date })
+    return renderDateBadge(date)
   }
 }
