+++ services/github/github-issue-detail.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { test, given } from 'sazerac'
-import { age } from '../color-formatters.js'
-import { formatDate, metric } from '../text-formatters.js'
+import { age, formatDate } from '../date.js'
+import { metric } from '../text-formatters.js'
 import { InvalidResponse } from '../index.js'
 import GithubIssueDetail from './github-issue-detail.service.js'
 import { issueStateColor, commentsColor } from './github-helpers.js'
