+++ services/github/github-release-date.service.js
@@ -1,8 +1,6 @@
-import dayjs from 'dayjs'
 import Joi from 'joi'
 import { pathParam, queryParam } from '../index.js'
-import { age } from '../color-formatters.js'
-import { formatDate } from '../text-formatters.js'
+import { renderDateBadge } from '../date.js'
 import { GithubAuthV3Service } from './github-auth-service.js'
 import { documentation, httpErrorsFor } from './github-helpers.js'
 
@@ -63,14 +61,6 @@ export default class GithubReleaseDate extends GithubAuthV3Service {
 
   static defaultBadgeData = { label: 'release date' }
 
-  static render({ date }) {
-    const releaseDate = dayjs(date)
-    return {
-      message: formatDate(releaseDate),
-      color: age(releaseDate),
-    }
-  }
-
   async fetch({ variant, user, repo }) {
     const url =
       variant === 'release-date'
@@ -86,10 +76,8 @@ export default class GithubReleaseDate extends GithubAuthV3Service {
   async handle({ variant, user, repo }, queryParams) {
     const body = await this.fetch({ variant, user, repo })
     if (Array.isArray(body)) {
-      return this.constructor.render({
-        date: body[0][queryParams.display_date],
-      })
+      return renderDateBadge(body[0][queryParams.display_date])
     }
-    return this.constructor.render({ date: body[queryParams.display_date] })
+    return renderDateBadge(body[queryParams.display_date])
   }
 }
