+++ services/open-vsx/open-vsx-release-date.service.js
@@ -1,6 +1,5 @@
 import { pathParams } from '../index.js'
-import { age } from '../color-formatters.js'
-import { formatDate } from '../text-formatters.js'
+import { renderDateBadge } from '../date.js'
 import { OpenVSXBase, description } from './open-vsx-base.js'
 
 export default class OpenVSXReleaseDate extends OpenVSXBase {
@@ -32,17 +31,8 @@ export default class OpenVSXReleaseDate extends OpenVSXBase {
 
   static defaultBadgeData = { label: 'release date' }
 
-  static render({ releaseDate }) {
-    return {
-      message: formatDate(releaseDate),
-      color: age(releaseDate),
-    }
-  }
-
   async handle({ namespace, extension }) {
     const { timestamp } = await this.fetch({ namespace, extension })
-    return this.constructor.render({
-      releaseDate: timestamp,
-    })
+    return renderDateBadge(timestamp)
   }
 }
