+++ services/sourceforge/sourceforge-last-commit.service.js
@@ -1,7 +1,6 @@
 import Joi from 'joi'
 import { BaseJsonService, pathParams } from '../index.js'
-import { formatDate } from '../text-formatters.js'
-import { age as ageColor } from '../color-formatters.js'
+import { renderDateBadge } from '../date.js'
 
 const schema = Joi.object({
   commits: Joi.array()
@@ -35,13 +34,6 @@ export default class SourceforgeLastCommit extends BaseJsonService {
 
   static defaultBadgeData = { label: 'last commit' }
 
-  static render({ commitDate }) {
-    return {
-      message: formatDate(new Date(commitDate)),
-      color: ageColor(new Date(commitDate)),
-    }
-  }
-
   async fetch({ project }) {
     return this._requestJson({
       url: `https://sourceforge.net/rest/p/${project}/git/commits`,
@@ -54,8 +46,6 @@ export default class SourceforgeLastCommit extends BaseJsonService {
 
   async handle({ project }) {
     const body = await this.fetch({ project })
-    return this.constructor.render({
-      commitDate: body.commits[0].committed_date,
-    })
+    return renderDateBadge(body.commits[0].committed_date)
   }
 }
