+++ services/text-formatters.js
@@ -6,9 +6,7 @@
  */
 
 import dayjs from 'dayjs'
-import calendar from 'dayjs/plugin/calendar.js'
 import relativeTime from 'dayjs/plugin/relativeTime.js'
-dayjs.extend(calendar)
 dayjs.extend(relativeTime)
 
 /**
@@ -165,39 +163,6 @@ function maybePluralize(singular, countable, plural) {
   }
 }
 
-/**
- * Returns a formatted date string without the year based on the value of input date param d.
- *
- * @param {Date | string | number | object } d - Input date in dayjs compatible format, date object, datestring, Unix timestamp etc.
- * @returns {string} Formatted date string
- */
-function formatDate(d) {
-  const date = dayjs(d)
-  const dateString = date.calendar(null, {
-    lastDay: '[yesterday]',
-    sameDay: '[today]',
-    lastWeek: '[last] dddd',
-    sameElse: 'MMMM YYYY',
-  })
-  // Trim current year from date string
-  return dateString.replace(` ${dayjs().year()}`, '').toLowerCase()
-}
-
-/**
- * Returns a relative date from the input timestamp.
- * For example, day after tomorrow's timestamp will return 'in 2 days'.
- *
- * @param {number | string} timestamp - Unix timestamp
- * @returns {string} Relative date from the unix timestamp
- */
-function formatRelativeDate(timestamp) {
-  const parsedDate = dayjs.unix(parseInt(timestamp, 10))
-  if (!parsedDate.isValid()) {
-    return 'invalid date'
-  }
-  return dayjs().to(parsedDate).toLowerCase()
-}
-
 export {
   starRating,
   currencyFromCode,
@@ -206,6 +171,4 @@ export {
   omitv,
   addv,
   maybePluralize,
-  formatDate,
-  formatRelativeDate,
 }
